Moved some interfaces from props to state.ts
