+++ src/webparts/trackMyTime/components/ITrackMyTimeState.ts
@@ -1,14 +1,79 @@
 
-import { ITrackMyTimeProps, IProject, IProjects, ITimeEntry } from './components/ITrackMyTimeProps';
+import { ITrackMyTimeProps } from './ITrackMyTimeProps';
+
+export interface IUser {
+  title: string;
+  initials?: string;  //Single person column
+  email?: string;  //Single person column
+  id: any;
+}
+
+export interface ILink {
+  description: string;
+  url: string;
+}
+
+export interface IUser {
+  title: string;
+  initials?: string;  //Single person column
+  email?: string;  //Single person column
+  id: any;
+}
+
+export interface ITimeEntry {
+  //Values that would come from Project item
+  titleProject: string;
+  category1?: string[];
+  category2?: string[];
+  leader?: IUser;  //Likely single person column
+  team?: IUser[];  //Likely multi person column
+
+  //Values that relate to project list item
+  sourceProject?: ILink; //Link back to the source project list item.
+
+  //Values specific to Time Entry
+  user: IUser;  //Single person column
+  startTime: any; //Time stamp
+  endTime: any; // Time stamp
+  duration: any; //Number
+
+  //Other settings and information
+  settings?: string;
+
+}
+
+export interface IProject {
+  //Values that would come from Project item
+  titleProject: string;
+
+  category1?: string[];
+  category2?: string[];
+  leader?: IUser;  //Likely single person column
+  team?: IUser[];  //Likely multi person column
+
+  //Values that relate to project list item
+  sourceProject?: ILink; //Link back to the source project list item.
+
+}
+
+export interface IProjects {
+  projects: IProject[];
+}
+
+export interface IProjectInfo {
+
+  master: IProjects[]; //Projects coming from the Projects list
+  user: IProjects[]; //Projects coming from TrackMyTime list
+  masterPriority: IProjects[]; //Projects visible based on settings
+  userPriority: IProjects[]; //Projects visible based on settings
+  current: IProjects[]; //Makes up the choices
+  
+}
 
 export interface ITrackMyTimeState {
   description: string;
 
-  projectsMaster: IProjects[]; //Projects coming from the Projects list
-  projectsUser: IProjects[]; //Projects coming from TrackMyTime list
-  projectsMasterPriority: IProjects[]; //Projects visible based on settings
-  projectsUserPriority: IProjects[]; //Projects visible based on settings
-  projectsCurrent: IProjects[]; //Makes up the choices
+  projects: IProjectInfo;
 
   projectMasterPriorityChoice: string; //Use to determine what projects float to top.... your most recent?  last day?
   projectUserPriorityChoice: string; //Use to determine what projects float to top.... your most recent?  last day?
